fix(datepicker): ExpressionChangedAfterItHasBeenCheckedError trackBy workaround
Uses trackBy in ngFor loops in the datepicker to avoid recreating DOM nodes
when navigating from one month to another.

This fixes the following issue: when using the keyboard to navigate from
one month to another, the focused element was destroyed during the change
detection of the keyboard event, which synchronously triggered the focusout
event in Chrome, leading to a change in the data model, and this was
causing the ExpressionChangedAfterItHasBeenCheckedError error.

Closes #2408